usb: misc: Set the IN and OUT EP to NULL
Currently in the probe function we are checking the IN and OUT
endpoint and if it is NULL then we are updating the endpoint
address but we are not set the IN and OUT EP to 0, hence update
of endpoing address is not happen in probe function and this result
incorrect mapping of endpoint address.

To avoid this set the IN and OUT EP to 0.

Change-Id: Id57dc559fff572381ae8439833535588413e24c0
Signed-off-by: Aniket Randive <arandive@codeaurora.org>